move port classification to ip with optional aggregate
